PSDto3D build instructions, and project cleanup
Added README.md with build instructions for PSDtoMaya and PSDtoFBX.  Cleaned up project files and build warnings.

Notes:
- Added a new README.md markdown file, the first thing shown on GitHub when browsing a repository.  Structure and formatting are inspired from the Unreal Engine source repository.  This includes detailed instructions for building PSDtoMaya, PSDtoFBX, and the installers.  It covers setup of a development environment, providing download location for third-party components, configuration of dependencies, and build instructions for Qt 5.15.2.
- Removed PSDtoMaya support for Maya 2017, 2018 and 2019, to simplify build and install around only the latest five versions.
- Removed the deps folder with pre-built Qt dependencies.  Instead PSDtoFBX now copies binaries from the Qt build location.  All required .dll files must now be present in the \bin subfolder of the Qt build.
- Renamed build configuration RelWithDebInfo_Standalone to RelWithDebInfo_FBX for clarity, and removed some unmaintained configurations.
- Fixed build warnings.  Added typecast from double to float as needed.  Added build option to suppress warnings about missing .pdb files in third-party libraries.

+++ psd2m_mesh_generator/src/mesh_generator/delaunay_mesh/delaunayAlgo.cpp
@@ -86,9 +86,9 @@ namespace mesh_generator
 	// assumes the interior is always "to the right" as the spline travels
 	bool IsEdgeInterior( Vector2F& p_prev, Vector2F& p_cur, Vector2F& p_next, Vector2F& p_check )
 	{
-		float a_prev = atan2( p_prev.x-p_cur.x, p_prev.y-p_cur.y );
-		float a_next = atan2( p_next.x-p_cur.x, p_next.y-p_cur.y );
-		float a_check = atan2( p_check.x-p_cur.x, p_check.y-p_cur.y );
+		float a_prev = (float)atan2( p_prev.x-p_cur.x, p_prev.y-p_cur.y );
+		float a_next = (float)atan2( p_next.x-p_cur.x, p_next.y-p_cur.y );
+		float a_check = (float)atan2( p_check.x-p_cur.x, p_check.y-p_cur.y );
 
 		a_next = a_next - a_prev;
 		a_check = a_check - a_prev;
@@ -564,7 +564,7 @@ namespace mesh_generator
 		}
 		sx /= c, sy /= c;
 		// solve for z:  sqrt(x^2 + y^2 + z^2)=1 ... square both sides, x^2 + y^2 + z^2 = 1 ... z = sqrt( 1 - (x^2 + y^2))
-		sz = sqrt( 1.0f - ((sx*sx) + (sy*sy)) );
+		sz = (float)sqrt( 1.0f - ((sx*sx) + (sy*sy)) );
 
 		int ix = (int)(127.5f + (127.5f*sx)), iy = (int)(127.5f + (127.5f*sy)), iz = (int)(127.5f + (127.5f*sz));
 		return NormalPix( (BYTE)ix, (BYTE)iy, (BYTE)iz );
